refactor: 프로필 카드의 스타일은 제외하고 info에 대한 스타일만 남겨 재사용 가능하도록 변경
