+++ src/components/common/profile-info/styles.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+const ProfileInfoGrid = styled.section`
+  display: grid;
+  gap: 18px 26px;
+  grid-template-columns: 42px 1fr;
+  grid-template-rows: 1fr 1fr;
+  width: 300px;
+`;
+
+export { ProfileInfoGrid };
