+++ src/components/common/profile-info/types.ts
@@ -1,0 +1,9 @@
+export interface ProfileInfo {
+  name: string;
+  nickname: string;
+  gender?: string;
+  birth: string;
+  profile_image_url: string;
+  mbti: string;
+  hobby: string[];
+}
