+++ src/components/pages/my-card/profile-card/styles.ts
@@ -30,18 +30,4 @@ const EditProfileButton = styled.button`
   box-shadow: 0 0 5px rgb(0 0 0 / 5%);
 `;
 
-const ProfileInfoGrid = styled.section`
-  display: grid;
-  gap: 18px 26px;
-  grid-template-columns: 42px 1fr;
-  grid-template-rows: 1fr 1fr;
-  width: 300px;
-`;
-
-export {
-  ProfileSection,
-  OpenedProfile,
-  ClosedProfile,
-  EditProfileButton,
-  ProfileInfoGrid,
-};
+export { ProfileSection, OpenedProfile, ClosedProfile, EditProfileButton };
