+++ src/components/pages/my-card/types.ts
@@ -1,13 +1,3 @@
-export interface ProfileInfo {
-  name: string;
-  nickname: string;
-  gender?: string;
-  birth: string;
-  profile_image_url: string;
-  mbti: string;
-  hobby: string[];
-}
-
 export type CardDirection = 'SENT' | 'RECEIVED';
 
 export type CardViewType = 'GRID' | 'LIST';
